Fixed issue with event perference matching
Previously, a single event was selected for each booking preference made
by the user. If this wasn't the correct event (eg. 'Formal' rather than
'Sunday Formal') then there would be an error when the booking was made.

Now, each matching event is returned, and when booking the events are
looped through until a booking is successfully made.